src/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
frozenlist==1.5.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
lightstreamer-client-lib==2.2.0
multidict==6.1.0
prometheus_client==0.21.1
propcache==0.2.1
yarl==1.18.3
